# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BetUZ Casino & Sports Betting API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  username: string;
  password: string;
  phone: string;
  referralCode?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface UserProfile {
  id: number;
  userId: string;
  username: string;
  phone: string;
  balance: number;
  firstName?: string;
  lastName?: string;
  country?: string;
  city?: string;
  isBlocked: boolean;
  isProfileComplete: boolean;
  totalDeposited: number;
  totalWagered: number;
  referralCode: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface UpdateProfileRequest {
  firstName?: string;
  lastName?: string;
  country?: string;
  city?: string;
  phone?: string;
}

export interface BalanceResponse {
  balance: number;
}

export interface ReferralInfo {
  referralCode: string;
  referredCount: number;
  qualifiedCount: number;
  bonusEarned: number;
  bonusAvailable: number;
}

export interface Transaction {
  id: number;
  type: string;
  amount: number;
  status: string;
  description: string;
  createdAt: string;
}

export interface TransactionList {
  transactions: Transaction[];
}

export type PlayGameRequestGameType =
  (typeof PlayGameRequestGameType)[keyof typeof PlayGameRequestGameType];

export const PlayGameRequestGameType = {
  slots: "slots",
  mines: "mines",
  crash: "crash",
  dice: "dice",
  wheel: "wheel",
  roulette: "roulette",
  blackjack: "blackjack",
  baccarat: "baccarat",
  keno: "keno",
  plinko: "plinko",
  hilo: "hilo",
  coinflip: "coinflip",
  tower: "tower",
  applefortune: "applefortune",
  crystal: "crystal",
  dragontiger: "dragontiger",
  sicbo: "sicbo",
  andarbahar: "andarbahar",
  teenpatii: "teenpatii",
  videopoker: "videopoker",
} as const;

export type PlayGameRequestGameData = { [key: string]: unknown };

export interface PlayGameRequest {
  gameType: PlayGameRequestGameType;
  betAmount: number;
  gameData?: PlayGameRequestGameData;
}

export type GameResultResult = { [key: string]: unknown };

export interface GameResult {
  won: boolean;
  winAmount: number;
  newBalance: number;
  multiplier: number;
  result?: GameResultResult;
  message: string;
}

export interface GameItem {
  id: string;
  name: string;
  nameUz: string;
  type: string;
  category: string;
  maxWin: string;
  isHot: boolean;
  isBest: boolean;
}

export interface GamesList {
  games: GameItem[];
}

export type DepositInitiateRequestMethod =
  (typeof DepositInitiateRequestMethod)[keyof typeof DepositInitiateRequestMethod];

export const DepositInitiateRequestMethod = {
  uzcard: "uzcard",
  humo: "humo",
  visa: "visa",
} as const;

export interface DepositInitiateRequest {
  method: DepositInitiateRequestMethod;
  amount: number;
}

export interface DepositResponse {
  depositId: number;
  cardNumber: string;
  cardHolder: string;
  method: string;
  amount: number;
}

export interface DepositConfirmRequest {
  depositId: number;
  receipt?: Blob;
}

export type WithdrawalRequestMethod =
  (typeof WithdrawalRequestMethod)[keyof typeof WithdrawalRequestMethod];

export const WithdrawalRequestMethod = {
  uzcard: "uzcard",
  humo: "humo",
  visa: "visa",
} as const;

export interface WithdrawalRequest {
  method: WithdrawalRequestMethod;
  amount: number;
  cardNumber: string;
}

export interface Odds {
  home: number;
  draw: number;
  away: number;
}

export interface Match {
  id: string;
  league: string;
  homeTeam: string;
  awayTeam: string;
  homeScore: number;
  awayScore: number;
  status: string;
  minute?: number;
  startTime: string;
  odds: Odds;
  isLive: boolean;
}

export interface MatchesList {
  live: Match[];
  upcoming: Match[];
}

export type SportsBetRequestBetType =
  (typeof SportsBetRequestBetType)[keyof typeof SportsBetRequestBetType];

export const SportsBetRequestBetType = {
  home: "home",
  draw: "draw",
  away: "away",
} as const;

export interface SportsBetRequest {
  matchId: string;
  betType: SportsBetRequestBetType;
  betAmount: number;
  odds: number;
}

export interface SupportMessageRequest {
  message?: string;
  file?: Blob;
}

export interface AdminVerifyRequest {
  code: string;
}

export interface AdminUserItem {
  id: number;
  userId: string;
  username: string;
  phone: string;
  balance: number;
  isBlocked: boolean;
  totalDeposited: number;
  createdAt: string;
}

export interface AdminUsersList {
  users: AdminUserItem[];
}

export type AdminBalanceRequestAction =
  (typeof AdminBalanceRequestAction)[keyof typeof AdminBalanceRequestAction];

export const AdminBalanceRequestAction = {
  add: "add",
  subtract: "subtract",
} as const;

export interface AdminBalanceRequest {
  userId: string;
  amount: number;
  action: AdminBalanceRequestAction;
  reason?: string;
}

export type AdminBlockRequestAction =
  (typeof AdminBlockRequestAction)[keyof typeof AdminBlockRequestAction];

export const AdminBlockRequestAction = {
  block: "block",
  unblock: "unblock",
} as const;

export interface AdminBlockRequest {
  userId: string;
  action: AdminBlockRequestAction;
}