# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BetUZ Casino & Sports Betting API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/verify:
    post:
      operationId: verifyAdmin
      tags: [admin]
      summary: Verify admin code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminVerifyRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/balance:
    post:
      operationId: adminUpdateBalance
      tags: [admin]
      summary: Add or subtract balance (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBalanceRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/block:
    post:
      operationId: adminBlockUser
      tags: [admin]
      summary: Block or unblock user (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBlockRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        phone:
          type: string
        referralCode:
          type: string
      required: [username, password, phone]
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [token, user]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        username:
          type: string
        phone:
          type: string
        balance:
          type: number
        firstName:
          type: string
        lastName:
          type: string
        country:
          type: string
        city:
          type: string
        isBlocked:
          type: boolean
        isProfileComplete:
          type: boolean
        totalDeposited:
          type: number
        totalWagered:
          type: number
        referralCode:
          type: string
        createdAt:
          type: string
      required: [id, userId, username, phone, balance, isBlocked, isProfileComplete, totalDeposited, totalWagered, referralCode, createdAt]
...
    AdminVerifyRequest:
      type: object
      properties:
        code:
          type: string
      required: [code]
...
    AdminBalanceRequest:
      type: object
      properties:
        userId:
          type: string
        amount:
          type: number
        action:
          type: string
          enum: [add, subtract]
        reason:
          type: string
      required: [userId, amount, action]
    AdminBlockRequest:
      type: object
      properties:
        userId:
          type: string
        action:
          type: string
          enum: [block, unblock]
      required: [userId, action]